[TASK] Optimize PageContentFetchingProcessor
This change adapts the newly introduced PageContentFetching
data processor, which reduces the number of SQL queries
for colPos entries that do not use the "slideMode" feature.

This means, that installations which use
- multiple colPos on one layout

will have 1 instead of "amountOfColPos" DB queries
for fetching all content of a page.

For colPos with a slide mode, the previous logic
still applies.

Resolves: #103906
Related: #103894
Releases: main
Change-Id: I5172fa427a69f279e88c27f9b97793036d94c533
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84387
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>