[BUGFIX] Reinitialize ContentObjectRenderer for pages
The menu generation must not reuse the page record from the current page
but must use the data from the linked page.

This got already fixed for #101883.
But that fix didn't take into account that ->createLink() might throw an
exception. This is now considered and the state is properly
re- initialized also on exceptions.

Resolves: #104335
Related: #103582
Related: #101883
Releases: main, 12.4
Change-Id: Ib4c1f1fba68a83dca32a9ffd7ee14c31cf60bab5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85175
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>