[FEATURE] Automatic frontend cache tagging
Introduce more sophisticated automatic cache tagging in multiple
places of the core.

Extbase:
* A Cache tag is added for the tablename when using findAll() and
  countAll().
* A Cache tag for the storage pid is added when using th findBy*()
  and __call() methods

ContentObject:
* A Cache Tag is added for each records rendered by
  ContentContentObject and RecordContentObject

TypoLink:
* A Cache Tag is added for linked pages. This is useful for scenarios
  like having a page with start and/or end time. When a page links to
  this kind of pages the link should reflect the correct state.
  Currently it could happen the target page is "hidden" but the page
  that linked to it was served from cache and the link was still
  active.

Resources:
* A Cache Tag is added when linking to a file. This is useful when a
  file is replaced or metadata are changed the.

This feature is guarded with toggle "frontend.cache.autoTagging" and
experimental for now: The feature is still incomplete and there should
be a couple more places that add tags. Additionally, some instances
may tend to add many tags, which may have a negative impact on
performance. Automatic tagging and cache clearing by tag is the way
to go, though. But we have to gather additional feedback from the
community to gain knowledge in order to determine loose ends and
bottlenecks. As such, we decided to bring this feature into the wild
now. The feature toggle is by default *disabled* for instances being
upgraded from earlier core versions, and *enabled* for newly
created instances based on TYPO3 v13.

Resolves: #104990
Related: #102422
Releases: main
Change-Id: I839ebc5bac7a066fb158a5213d419956937b44dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85778
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Sascha Nowak <typo3@saschanowak.me>
Reviewed-by: Sascha Nowak <typo3@saschanowak.me>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>