[FEATURE] Introduce ContentArea
A new ContentArea object is introduced and contains
next to the area specific information like the
column identifier, the colPos, the slideMode, etc.
also the corresponding records (content elements).

The new object is added as a Closure to the page
layout and initializes on access only => resolving
the content elements only when needed.

Additionally, the content areas are packed into
the new ContentAreaCollection, which allows access
via the known syntax in fluid by implementing the
PSR-11 ContainerInterface.

The ContentArea is based on the new ContentAreaInterface.
The interface allows third-party tools like EXT:container,
to create area objects for nested areas.

The PAGEVIEW cObject automatically exposes this
collection to the view using the varibale name
defined via the "contentAs" option or by falling
back to "content". This further reduces the
necessary TypoScript for rendering a page
and most likely make the PageContentFetchingProcessor
obsolet.

Using the ContentArea it's now also possible for
each content element to know the exact context
in which it should be rendered. This allows
to vary in used Layouts and Partials, depending
on e.g. the colPos the element is rendered in
=> skip the header in case the element is
rendered in the sidebar area.

Side note: An identifier for a column will be
mandatory in v15. For now a fallback is applied
(using a hash of the layout name and the colPos).

Resolves: #104974
Releases: main
Change-Id: I0fac3cd6d14b497ae25cc27793c4cc937f5bb2d3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86104
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Matthias Vogel <typo3@kanti.de>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Matthias Vogel <typo3@kanti.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>