[TASK] Use frontend sub-request for `ShowImageControllerTest`
Using method mocks to define what they should return to check
expectations only tests that the code flow does not change it.
This could be compared to a pure pipeline flow test to see if
water added on one end reaches the other end with same volume.

Tests based on this technique are useful in many cases,
but it is not the right way to verify and ensure that
something really returns what the normal code flow would
return. If the expectation in a test is not adjusted along
with the mocked (faked) return value such a test would
never reveal that the expecation does not match the real
world.

This change refactors `ShowImageControllerTest` using
as instance setup and frontend sub-requests
to test the "eID" handler implementation.

That reveales the expectation of having a leading
slash for the image path is incorrect. That is already
the case in real world instances.

Note it makes fully sense that such an early eID
handler using the FAL systems could not prefix images based
on current workflow and setup of the stack.

Resolves: #105006
Releases: main
Change-Id: I7eb987fe5cfbaa001fc62e12a4d8052e328e882b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86143
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>