[BUGFIX] Render hreflang properly in TypoScript getData
"hreflang" can be set manually in the site config, but
if empty, the locale is used by default.

Before this patch, the fallback to "locale:full"
prevented to actually retrieve the specific
"hreflang" attribute (available via
`siteLanguage->toArray()['hreflang']`).

With this patch, the override is removed, retrieving
the actual value if set, or falling back thanks
to the `toArray` fallback implementation:

```
  'hreflang' => $this->hreflang ?: $this->locale->getName(),
```

As a drive-by, a not-needed null-coalescing on the
return value is removed as well.

Resolves: #104987
Related: #102178
Related: #101160
Related: #95798
Related: #99908
Releases: main, 13.4
Change-Id: I09db38150aa86a75953dedbce0271ca93da7e15c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91085
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>