[BUGFIX] Reference correct time fields in cache calculation
The method `CacheLifetimeCalculator::getFirstTimeValueForRecord()`
collects the configured time fields (starttime, endtime) by their actual
field names, which can be different, depending on the TCA.

However, when calculating the value the TCA setting names were used as
field names, which happens to always match within the TYPO3 Core, but
may not in 3rd party extensions.

The method is fixed to use the correct field names to access record
properties.

Resolves: #91693
Releases: main, 13.4, 12.4
Change-Id: I163cb9e2ef8c54cef7f71da044df48e189062145
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86864
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>