[BUGFIX] Fix site settings hash calculation
The site-constants cache key in method
`addDefaultTypoScriptConstantsFromSite` of
`SysTemplateTreeBuilder` uses `json_encode()`
in order to build a hash of the site settings object.

With #104858 site settings were expanded to
allow setting-key access via `$settings->get('foo.bar')`
for defined settings and missed to adapt the `jsonSerialize()`
method to continue to return the settings tree instead of the
reduced settings map (which is limited to
settings defined in `settings.definitions.yaml`).

This leads to broken TS caches with multiple sites:
Settings from one site spill over to other sites,
depending on which one was called first after cache clear.

A regression test replays the scenario to proof correctness
of the code change.

Releases: main, 13.4
Resolves: #105760
Related: #104858
Change-Id: Ied715c25265abb98681b11c13468679c0b23528a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87557
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Achim Fritz <af@achimfritz.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Achim Fritz <af@achimfritz.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>