[BUGFIX] Check the shortcut target page UID to identify its ACT and CUR state
Introduced in #99838, the uid in _SHORTCUT_ORIGINAL_PAGE_UID was changed to the logically correct value, but it broke the behavior for menu itemStates.
This fix checks if a shortcut target was resolved (which is the case when _SHORTCUT_ORIGINAL_PAGE_UID is set) and the uses the uid of the resolved page instead of the uid of the shortcut page again.

Resolves: #106210
Resolves: #97186
Releases: main, 13.4
Change-Id: I1388b0c3f8fcb465b87fc5fdc783b03d7cbd4c09
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88740
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>