[BUGFIX] Avoid PHP bug when a link was not created via HtmlWorker
When using HtmlWorker (e.g. in Backend Login's system news),
there was a PHP error in DomDocument.

When a link was not possible to generate (example: a page
was deleted) then a "UnableToLinkException" is thrown
and just the contents of the <a> tag (incl. spans, strong etc)
are rendered.

However, as it turns out, the inner workings of replacing
this in DomElement (as done in HtmlWorker, whereas in
lib.parseFunc this is purely done with a string/array magic)
"in-place" causes a memory issue resulting in killing
PHP on ones' server (100% system load).

The solution is to clone the child nodes in DomDocument, and
append them before the element, then removing the actual
<a> tag ($element) in this case.

Resolves: #106425
Releases: main, 13.4
Change-Id: I4082d0e45a83c97b9fc8b00fbeebc3cd548d7d0c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88737
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>