[BUGFIX] Ignore 'logintype' for cHash validation
'logintype' is a core parameter and may be
attached to any URL (e.g. for logintype=logout).
Moreover, alternative authentication services may
need to add a logintype=login to an existing URL.

Since cHash is the mechanics for avoidance of
caching flooding, ignoring the logintype parameter
will not cause any issue in that regard.

Since #95297 the ['cacheHash']['enforceValidation']
configuration option is enabled as part of the
factory configuration by default.
If this is enabled, any attempt to use the logintype
parameter will cause an error.

Resolves: #106835
Related: #95297
Releases: main, 13.4, 12.4
Change-Id: I54e1ce4560c80e736aa3864ac413080d6a1652c3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89657
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>