[TASK] Require and use dependency injection in database services
Allow for clean TCA Schema usage, whose factory requires the
symfony container to be used, therefore we now limit ourselves
to only use the database connection on a fully booted system
and not in a failsafe environment.

This allows to drop workarounds we added with #106867 and
allows to rely on dependency injection and avoid `GU::makeInstance`
in database related services or restrictions.

The consequence is that all usages of the connection pool inside
the install tool now ensures to boot into the symfony container
in order to provide the requirements that are now needed for the
connection pool to run. This is a continuation of the efforts
started in #110128 in order to reduce failsafe provided services
into those that are actually needed for failsafe operation (which
is to recover from a broken installation).

Resolves: #110163
Related: #96291
Related: #106867
Related: #110128
Releases: main
Change-Id: Id9210f13c4f82065288611a2cc1d9c58abc51bb4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89956
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>