[TASK] Model TS "register" as request attribute
The frontend TypoScript related "register" functionality is one
of the very old structures that are pretty hard to understand
nowadays and don't find that many usages anymore.
It is one of the construct we do have to touch with current
TSFE works, but we shouldn't break it either for no good
reason.

The "register" is a combination of a key/value store with a
stack. The patch models this as two classes, attaches an
instance of the stack as request attribute 'frontend.register.stack',
adapts usages and removes the two TSFE properties $register and
$registerStack.

The change needs a patch in typo3/testing-framework as well,
so we raise it:

> composer u typo3/testing-framework

Resolves: #107625
Releases: main
Change-Id: If79982963baf514f4a2e2d5571d7152f7f0bd7fd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90945
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>