[!!!][TASK] Streamline PSR-7 response header handling
This patch standardizes and centralizes PSR-7 response header
processing across Core and Extbase. Previously, Extbase only kept
the last value when a single plugin provided multiple header values,
which made it impossible to emit multiple headers with the same name
such as Set-Cookie. The Core, on the other hand, supported multiple
Set-Cookie headers via an allowlist but merged all other headers with
multiple values into a single comma-separated string. According to
RFC 9110 [1], this is only valid if the header explicitly allows
comma-separated lists.

With this change, multiple header values are preserved if created by
a single extbase plugin, resulting in multiple header lines. A single
value is emitted as a single-line header. This replaces the prior
allowlist and inconsistent behavior with a unified approach.

The structural win with this change is that extbase no longer
triggers header() calls on its own.

The solution does not deal with another detail: Multiple extbase
plugins can not set multiple header lines (e.g. two plugins can not
both set own 'set-cookie'), the last plugin will override the former.
The "last one wins" approach is kept to retain b/w compatibility and
turning this around would break the "i want to overwrite" scenario.
The issue stems from a limitation based on the situation that extbase
plugins return Response objects, which does not allow a distinction
between 'overriding' and 'appending' as such, to notify 'merger' code
in RequestHandler or ContentObjectRenderer on how to act. This is one
of the places where the decision to return PSR-7 Response objects within
extbase bites us: This should be a "content part" or something like that
instead to allow proper merging data in an upper layer. Instances
facing this issue must for now deal with this using a middleware
(probably located *after* "typo3/cms-core/response-propagation" to not
catch the response in case one plugin throws a 3xx/4xx response) that
merges and sets a header compiplation on the Response object.

The new behavior supports additional use cases such as WWW-Authenticate
[2], Link [3] and xkey [4] headers, which are also supported by other
PHP frameworks including Symfony [5], Laravel, CakePHP [6], CodeIgniter
[7] and Yii [8].

[1] https://www.rfc-editor.org/rfc/rfc9110.html
[2] https://developer.mozilla.org/docs/Web/HTTP/Headers/WWW-Authenticate
[3] https://httpwg.org/specs/rfc8288.html
[4] https://github.com/varnish/varnish-modules/blob/master/src/vmod_xkey.vcc
[5] https://github.com/symfony/symfony/blob/e971be71d94871b316d2b11a7ba166d6229af450/src/Symfony/Component/HttpFoundation/Response.php#L335-L360
[6] https://github.com/cakephp/cakephp/blob/e05df58b3a5a1f52d64d3dad8a235337c0d52b12/src/Http/ResponseEmitter.php#L178
[7] https://github.com/codeigniter4/CodeIgniter4/blob/9972ca0fec7e919cbf39157d5c95141baab5cad2/system/HTTP/ResponseTrait.php#L405
[8] https://github.com/yiisoft/yii2/blob/976eac5443b2b6cee8263a4f2bdbf2e6c360102c/framework/web/Response.php#L366

Resolves: #107324
Resolves: #99032
Resolves: #103685
Releases: main
Change-Id: I80ff70cc48597a686540ef56e7afedff4378b6ac
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90471
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>