[BUGFIX] Allow Fluid access to SiteLanguage configuration
Accessing the `configuration` property of a `SiteLanguage`
object in Fluid was not straight-forward, because its
member properties are protected, and the `toArray()`
method is not accessible to Fluid to resolve the array
structure.

Similar to `Site`, the `SiteLanguage` object
can now access all configuration via a getter.

This allows to also access any configured custom
fields there with notation like:

```
Language: {siteLanguage.configuration.navigationTitle}
Customfield: {siteLanguage.configuration.customField}
```

(The `{siteLanguage}` variable would for example be
fetched with the SiteLanguageProcessor Dataprocessor,
or could also be retrieved from the SiteProcessor
in its sub-structure.)

Resolves: #107524
Releases: main, 13.4
Change-Id: Idd9897d60173ab1c1e3ccca20ae16195c4f515d3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90961
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>