[TASK] Move FrontendTypoScriptController->contentType
This patch itself is pretty straight forward, but needs
a bigger explanation on why we do this detail now and
why we solve it that way.

The entire story boils down to ContentObjectRenderer only
returning strings for single parts of a generated page.
This wasn't even a good idea back in the old days and
materialized in various odd constructs that tend to park
additional state elsewhere: When a content object is
rendered, it often not only wants to create some content
string that is part of the final body stream at the
current position of the generated body. It also often
wants to do additional stuff like registering some JS
in the head part of the HTML body, or some CSS, it may
want to change the HTML body tag, it may want to set
a different HTTP Response Content-Type or other header
data. All that kind of stuff: Influence the final HTTP
Response, and/or various different parts of the HTML body
stream.

But ContentObjectRenderer only allows a single cObj to return
strings. And that string is then simply concatenated
to what has been generated already. To still register stuff
like additional header files, the PageRenderer singleton has
been created to gather such body stream related things. In
addition to that, things like TSFE->contentType exists to carry
a HTTP Response detail. On top of that, various Request attributes
nowadays carry things like cache tags and cache lifetimes. And
we also have a lot of state in ContentObjectRenderer itself.
Last but not least, the frontend has the concept of uncached
elements ("USER_INT" / "COA_INT"), which park state in TSFE,
relies on a funny placeholder mechanism within the generated
body string, plus requires ContentObjectRenderer and other
instances to be serialized because its state at this point in
the rendering chain needs to be "frozen" and put into cache.
The Extbase bootstrap which fiddles a lot with Response shows
drawbacks of this approach quite well, and the 'uncached body
part' handling in TSFE is so brittle that we barely touched
it over the years, even though it has various issues and
details that don't work well in "INT" elements.

We will have to substitute the string return of single content
objects with a data object that can carry more in the future.
If solved in a good way, this will remove the need for the
stateful PageRenderer, various request attributes and more,
and we'll end up with a central place that compiles initial
Response.

We however don't know details yet and the topic is hard to
get right since we have to think about many things at once,
for instance Fluid has to be involved as well. But there are
things that have to happen as preparation: TSFE needs to vanish
to drop the cross dependency between TSFE and other things,
especially COR. Also COR needs to be turned into a service
loosing its state, which will improve "INT" handling already
and should allow getting rid of __wakeup() and __sleep().

This patch moves one further details away from TSFE: $contentType
by adding it to the already established (internal) PageParts
request attribute 'fronend.page.parts' instead. This class will
carry state for now and can be seen as a temporary solution that
represents major parts of technical debt we'll have to model
in a good way when surrounding stuff has been refactored long
enough to allow major changes.

Resolves: #107641
Releases: main
Change-Id: If569dab11eb2157f9c4cf9319c8dcab82771585f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90974
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>